Moved custom plugins into separate javascript modules.
